requirements.txt
# Core dependencies
ctransformers>=0.2.26
langchain>=0.0.208
langchain-community>=0.0.1
llama-cpp-python>=0.2.0
chromadb>=0.4.0
rich>=13.0.0
astunparse>=1.6.3
typer>=0.9.0
pytest>=7.0.0

# For new LLM integration
pathlib>=1.0.1

# Fix dependency warnings by pinning compatible versions
urllib3<2.0.0,>=1.26.0
charset-normalizer<3.0.0,>=2.0.0
requests>=2.28.0,<3.0.0
